requirements.txt
beautifulsoup4
Python-dotenv
markdownify
feedparser
requests
litellm
pandas